Overhaul the addition of silence to recordings, making it much less timing dependent.
Details: instead of tracking the time between packets received for a given user, and adding additional silence if a long gap occurs (which was prone to jitter since the timing was done by Instant::now(), now I keep a HashSet of all users who we've ever received a VoiceState from; on every VoiceTick, we duplicate the set and remove the users who were in VoiceTick.speaking. The leftovers are everyone who wasn't speaking that frame. It should be a superset of VoiceTick.silent, including those users, but also users who have been silent long enough to stop receiving packets from, or even users who have disconnected. All of these users get a packet's worth of silent samples added to their encoding buffer, meaning they're always kept up to real time even if we have no data from them. This means that all users who ever showed up in the call should have almost exactly the same length file. The one possibility for jitter is a late addition, where we do use an Instant to calculate a Duration since the beginning of recording and add a block of silence. However, the end of the files should be synchronized, so even if there is very slight jitter in the number of silence samples at the beginning, all users' recordings can be synced by aligning the end of the tracks.

Also: fixed duplicate VoiceReceivers on extra joins, made do_join() return the command sender, and removed the RecordingManager wrapper.